WIP Fix generali per upload e download files

+++ backend/PolarDrive.Data/DTOs/ClientConsentDTO.cs
@@ -8,7 +8,7 @@ public class ClientConsentDTO
     public string CompanyVatNumber { get; set; } = string.Empty;
     public string VehicleVIN { get; set; } = string.Empty;
     public string UploadDate { get; set; } = string.Empty;
-    public int ZipFileSize { get; set; } 
+    public long ZipFileSize { get; set; } 
     public string ConsentHash { get; set; } = string.Empty;
     public string ConsentType { get; set; } = string.Empty;
     public string? Notes { get; set; }
